+++ sigstore/_cli.py
@@ -24,15 +24,12 @@ from textwrap import dedent
 from typing import Optional, TextIO, Union, cast
 
 from cryptography.x509 import load_pem_x509_certificates
+from sigstore_protobuf_specs.dev.sigstore.bundle.v1 import Bundle
 
 from sigstore import __version__
 from sigstore._internal.ctfe import CTKeyring
 from sigstore._internal.fulcio.client import DEFAULT_FULCIO_URL, FulcioClient
-from sigstore._internal.rekor.client import (
-    DEFAULT_REKOR_URL,
-    RekorBundle,
-    RekorClient,
-)
+from sigstore._internal.rekor.client import DEFAULT_REKOR_URL, RekorClient
 from sigstore._internal.tuf import TrustUpdater
 from sigstore.oidc import (
     DEFAULT_OAUTH_ISSUER_URL,
@@ -179,21 +176,11 @@ def _add_shared_verify_input_options(group: argparse._ArgumentGroup) -> None:
         help="The signature to verify against; not used with multiple inputs",
     )
     group.add_argument(
-        "--rekor-bundle",
-        metavar="FILE",
-        type=Path,
-        default=os.getenv("SIGSTORE_REKOR_BUNDLE"),
-        help="The offline Rekor bundle to verify with; not used with multiple inputs",
-    )
-    group.add_argument(
         "--bundle",
         metavar="FILE",
         type=Path,
         default=os.getenv("SIGSTORE_BUNDLE"),
-        help=(
-            "The Sigstore bundle to verify with; not used with multiple inputs; this option is "
-            "experimental and may change between releases until stabilized"
-        ),
+        help=("The Sigstore bundle to verify with; not used with multiple inputs"),
     )
     group.add_argument(
         "files",
@@ -214,10 +201,10 @@ def _add_shared_verification_options(group: argparse._ArgumentGroup) -> None:
         required=True,
     )
     group.add_argument(
-        "--require-rekor-offline",
+        "--offline",
         action="store_true",
-        default=_boolify_env("SIGSTORE_REQUIRE_REKOR_OFFLINE"),
-        help="Require offline Rekor verification with a bundle; implied by --rekor-bundle",
+        default=_boolify_env("SIGSTORE_OFFLINE"),
+        help="Perform offline verification; requires a Sigstore bundle",
     )
 
 
@@ -340,24 +327,13 @@ def _parser() -> argparse.ArgumentParser:
         ),
     )
     output_options.add_argument(
-        "--rekor-bundle",
-        "--output-rekor-bundle",
-        metavar="FILE",
-        type=Path,
-        default=os.getenv("SIGSTORE_OUTPUT_BUNDLE"),
-        help=(
-            "Write a single offline Rekor bundle to the given file; does not work with "
-            "multiple input files"
-        ),
-    )
-    output_options.add_argument(
         "--bundle",
         metavar="FILE",
         type=Path,
         default=os.getenv("SIGSTORE_BUNDLE"),
         help=(
             "Write a single Sigstore bundle to the given file; does not work with multiple input "
-            "files; this option is experimental and may change between releases until stabilized"
+            "files"
         ),
     )
     output_options.add_argument(
@@ -577,14 +553,6 @@ def main() -> None:
 
 
 def _sign(args: argparse.Namespace) -> None:
-    # `--rekor-bundle` is a temporary option, pending stabilization of the
-    # Sigstore bundle format.
-    if args.rekor_bundle:
-        logger.warning(
-            "--rekor-bundle is a temporary format, and will be removed in an "
-            "upcoming release of sigstore-python in favor of Sigstore-style bundles"
-        )
-
     if args.bundle:
         logger.warning(
             "--bundle support is experimental; the behaviour of this flag may change "
@@ -597,20 +565,13 @@ def _sign(args: argparse.Namespace) -> None:
             "between releases until stabilized."
         )
 
-    # `--no-default-files` has no effect on `--{signature,certificate,rekor-bundle,bundle}`,
+    # `--no-default-files` has no effect on `--{signature,certificate,bundle}`,
     # but we forbid it because it indicates user confusion.
-    if args.no_default_files and (
-        args.signature or args.certificate or args.rekor_bundle or args.bundle
-    ):
+    if args.no_default_files and (args.signature or args.certificate or args.bundle):
         args._parser.error(
             "--no-default-files may not be combined with --signature, "
-            "--certificate, --rekor-bundle, or --bundle",
+            "--certificate, or --bundle",
         )
-
-    # Similarly forbid `--rekor-bundle` with `--bundle`, since it again indicates
-    # user confusion around outputs.
-    if args.rekor_bundle and args.bundle:
-        args._parser.error("--rekor-bundle may not be combined with --bundle")
 
     # Fail if `--bundle` and `--no-bundle` are both specified.
     if args.bundle and args.no_bundle:
@@ -618,11 +579,9 @@ def _sign(args: argparse.Namespace) -> None:
 
     # Fail if `--signature` or `--certificate` is specified *and* we have more
     # than one input.
-    if (args.signature or args.certificate or args.rekor_bundle) and len(
-        args.files
-    ) > 1:
+    if (args.signature or args.certificate) and len(args.files) > 1:
         args._parser.error(
-            "Error: --signature, --certificate, and --rekor-bundle can't be used "
+            "Error: --signature and --certificate can't be used "
             "with explicit outputs for multiple inputs",
         )
 
@@ -633,22 +592,14 @@ def _sign(args: argparse.Namespace) -> None:
         if not file.is_file():
             args._parser.error(f"Input must be a file: {file}")
 
-        sig, cert, rekor_bundle, bundle = (
+        sig, cert, bundle = (
             args.signature,
             args.certificate,
-            args.rekor_bundle,
             args.bundle,
         )
-        if (
-            not sig
-            and not cert
-            and not rekor_bundle
-            and not bundle
-            and not args.no_default_files
-        ):
+        if not sig and not cert and not bundle and not args.no_default_files:
             sig = file.parent / f"{file.name}.sig"
             cert = file.parent / f"{file.name}.crt"
-            rekor_bundle = file.parent / f"{file.name}.rekor"
             if not args.no_bundle:
                 bundle = file.parent / f"{file.name}.sigstore"
 
@@ -658,8 +609,6 @@ def _sign(args: argparse.Namespace) -> None:
                 extants.append(str(sig))
             if cert and cert.exists():
                 extants.append(str(cert))
-            if rekor_bundle and rekor_bundle.exists():
-                extants.append(str(rekor_bundle))
             if bundle and bundle.exists():
                 extants.append(str(bundle))
 
@@ -672,7 +621,6 @@ def _sign(args: argparse.Namespace) -> None:
         output_map[file] = {
             "cert": cert,
             "sig": sig,
-            "rekor_bundle": rekor_bundle,
             "bundle": bundle,
         }
 
@@ -739,12 +687,6 @@ def _sign(args: argparse.Namespace) -> None:
                 print(result.cert_pem, file=io)
             print(f"Certificate written to {outputs['cert']}")
 
-        if outputs["rekor_bundle"] is not None:
-            with outputs["rekor_bundle"].open(mode="w") as io:
-                rekor_bundle = RekorBundle.from_entry(result.log_entry)
-                print(rekor_bundle.json(by_alias=True), file=io)
-            print(f"Rekor bundle written to {outputs['rekor_bundle']}")
-
         if outputs["bundle"] is not None:
             with outputs["bundle"].open(mode="w") as io:
                 print(result._to_bundle().to_json(), file=io)
@@ -762,30 +704,17 @@ def _collect_verification_state(
     purposes) and `materials` is the `VerificationMaterials` to verify with.
     """
 
-    # TODO: Allow --bundle during verification. Until then, error.
-    if args.bundle:
-        args._parser.error("--bundle is not supported during verification yet")
-
-    # `--rekor-bundle` is a temporary option, pending stabilization of the
-    # Sigstore bundle format.
-    if args.rekor_bundle:
-        logger.warning(
-            "--rekor-bundle is a temporary format, and will be removed in an "
-            "upcoming release of sigstore-python in favor of Sigstore-style bundles"
-        )
-
-    # The presence of --rekor-bundle implies --require-rekor-offline.
-    args.require_rekor_offline = args.require_rekor_offline or args.rekor_bundle
-
-    # Fail if --certificate, --signature, or --rekor-bundle is specified and we
+    # Fail if --certificate, --signature, or --bundle is specified and we
     # have more than one input.
-    if (args.certificate or args.signature or args.rekor_bundle) and len(
-        args.files
-    ) > 1:
+    if (args.certificate or args.signature or args.bundle) and len(args.files) > 1:
         args._parser.error(
-            "--certificate, --signature, and --rekor-bundle can only be used "
+            "--certificate, --signature, or --bundle can only be used "
             "with a single input file"
         )
+
+    # Fail if `--certificate` or `--signature` is used with `--bundle`.
+    if args.bundle and (args.certificate or args.signature):
+        args._parser.error("--bundle cannot be used with --certificate or --signature")
 
     # The converse of `sign`: we build up an expected input map and check
     # that we have everything so that we can fail early.
@@ -794,31 +723,37 @@ def _collect_verification_state(
         if not file.is_file():
             args._parser.error(f"Input must be a file: {file}")
 
-        sig, cert, bundle = args.signature, args.certificate, args.rekor_bundle
+        sig, cert, bundle = (
+            args.signature,
+            args.certificate,
+            args.bundle,
+        )
         if sig is None:
             sig = file.parent / f"{file.name}.sig"
         if cert is None:
             cert = file.parent / f"{file.name}.crt"
         if bundle is None:
-            bundle = file.parent / f"{file.name}.rekor"
+            bundle = file.parent / f"{file.name}.sigstore"
 
         missing = []
-        if not sig.is_file():
-            missing.append(str(sig))
-        if not cert.is_file():
-            missing.append(str(cert))
-        if not bundle.is_file() and args.require_rekor_offline:
-            # NOTE: We only produce errors on missing bundle files
-            # if the user has explicitly requested offline-only verification.
-            # Otherwise, we fall back on online verification.
-            missing.append(str(bundle))
+        if args.signature or args.certificate:
+            if not sig.is_file():
+                missing.append(str(sig))
+            if not cert.is_file():
+                missing.append(str(cert))
+            input_map[file] = {"cert": cert, "sig": sig}
+        else:
+            # If a user hasn't explicitly supplied `--signature`, `--certificate` or
+            # `--rekor-bundle`, we expect a bundle either supplied via `--bundle` or with the
+            # default `{input}.sigstore` name.
+            if not bundle.is_file():
+                missing.append(str(bundle))
+            input_map[file] = {"bundle": bundle}
 
         if missing:
             args._parser.error(
                 f"Missing verification materials for {(file)}: {', '.join(missing)}"
             )
-
-        input_map[file] = {"cert": cert, "sig": sig, "bundle": bundle}
 
     if args.staging:
         logger.debug("verify: staging instances requested")
@@ -856,34 +791,43 @@ def _collect_verification_state(
 
     all_materials = []
     for file, inputs in input_map.items():
-        # Load the signing certificate
-        logger.debug(f"Using certificate from: {inputs['cert']}")
-        cert_pem = inputs["cert"].read_text()
+        cert_pem: str
+        signature: bytes
+        entry: LogEntry | None = None
+        if "bundle" in inputs:
+            # Load the bundle
+            logger.debug(f"Using bundle from: {inputs['bundle']}")
 
-        # Load the signature
-        logger.debug(f"Using signature from: {inputs['sig']}")
-        b64_signature = inputs["sig"].read_text()
+            bundle_bytes = inputs["bundle"].read_bytes()
+            bundle = Bundle().from_json(bundle_bytes)
 
-        entry: Optional[LogEntry] = None
-        if inputs["bundle"].is_file():
-            logger.debug(f"Using offline Rekor bundle from: {inputs['bundle']}")
-            bundle = RekorBundle.parse_file(inputs["bundle"])
-            entry = bundle.to_entry()
+            with file.open(mode="rb", buffering=0) as io:
+                materials = VerificationMaterials.from_bundle(
+                    input_=io, bundle=bundle, offline=args.offline
+                )
+        else:
+            # Load the signing certificate
+            logger.debug(f"Using certificate from: {inputs['cert']}")
+            cert_pem = inputs["cert"].read_text()
+
+            # Load the signature
+            logger.debug(f"Using signature from: {inputs['sig']}")
+            b64_signature = inputs["sig"].read_text()
+            signature = base64.b64decode(b64_signature)
+
+            with file.open(mode="rb", buffering=0) as io:
+                materials = VerificationMaterials(
+                    input_=io,
+                    cert_pem=cert_pem,
+                    signature=signature,
+                    rekor_entry=entry,
+                    offline=args.offline,
+                )
 
         logger.debug(f"Verifying contents from: {file}")
 
         with file.open(mode="rb", buffering=0) as io:
-            all_materials.append(
-                (
-                    file,
-                    VerificationMaterials(
-                        input_=io,
-                        cert_pem=cert_pem,
-                        signature=base64.b64decode(b64_signature),
-                        offline_rekor_entry=entry,
-                    ),
-                )
-            )
+            all_materials.append((file, materials))
 
     return (verifier, all_materials)
 
@@ -891,7 +835,7 @@ def _collect_verification_state(
 def _verify_identity(args: argparse.Namespace) -> None:
     verifier, files_with_materials = _collect_verification_state(args)
 
-    for (file, materials) in files_with_materials:
+    for file, materials in files_with_materials:
         policy_ = policy.Identity(
             identity=args.cert_identity,
             issuer=args.cert_oidc_issuer,
@@ -990,7 +934,7 @@ def _verify_github(args: argparse.Namespace) -> None:
     policy_ = policy.AllOf(inner_policies)
 
     verifier, files_with_materials = _collect_verification_state(args)
-    for (file, materials) in files_with_materials:
+    for file, materials in files_with_materials:
         result = verifier.verify(materials=materials, policy=policy_)
 
         if result: